superclass for browseselection
