+++ commands/eventconfig.py
@@ -7,7 +7,7 @@ from discord.ext import commands
 import sqlite3
 
 from commands.interractions.playerconfig.memberconfig import PlayerConfig
-from commands.interractions.playerconfig.removememberconfig import BrowseSelection
+from commands.interractions.playerconfig.removememberconfig import RemoveMemberBrowseSelection
 from commands.utils.utils import haspermissions, tablify
 from discord.utils import escape_mentions
 from typing import Union
@@ -23,7 +23,7 @@ class Eventconfigurations(commands.Cog):
         self.databasepath = "./eventconfigurations.db"
     @commands.command(name="test")
     async def test(self, ctx):
-        view = BrowseSelection(["randomman32", "remastered151"], self.databasepath)
+        view = RemoveMemberBrowseSelection(["randomman32", "remastered151"], self.databasepath)
         await ctx.send('Pick your favourite colour:', view=view)
         await view.wait()
 
@@ -335,7 +335,7 @@ class Eventconfigurations(commands.Cog):
             cur.execute("SELECT playername FROM memberconfig WHERE guildid=?", (ctx.guild.id,))
             members = [row[0] for row in cur.fetchall()]
         if members:
-            view = BrowseSelection(members, self.databasepath, ctx)
+            view = RemoveMemberBrowseSelection(members, self.databasepath, ctx)
             await ctx.send('Pick your favourite colour:', view=view)
         else:
             await ctx.send("no members registered for playerconfig.")
